Refactor ComputedSeed logic for SeedVerifyView Move formatOrdinal from helpers to ComputedSeed
